Fix float type identifier

+++ core/src/main/java/galois/llm/query/utils/QueryUtils.java
@@ -132,7 +132,7 @@ public class QueryUtils {
 
     private static String getJsonSchemaTypeFromDBType(String type) {
         return switch (type.toLowerCase()) {
-            case "double precision", "real", "float" -> "number";
+            case "double precision", "real", "float8" -> "number";
             case "integer", "long" -> "integer";
             default -> "string";
         };
